+++ docs/CHANGELOG.md
@@ -3,6 +3,13 @@ RookScript (C) Black Rook Software
 by Matt Tropiano et al. (see AUTHORS.txt)
 
 
+Changed in [NOW]
+-----------------
+
+- `Added` ScriptValue.listApply() functions. 
+- `Changed` Better reflection with creation from lists and application to lists/arrays.
+
+
 Changed in 1.9.1
 ----------------